Extracted DeleteNoteMulti::task to deleteNoteMulti
